flight-delays/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.1.0
httpx==0.26.0
apscheduler==3.10.4
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2
python-multipart==0.0.6
tqdm==4.66.1
